+++ tests/test_generate.py
@@ -25,27 +25,30 @@ class CheckImplied(unittest.TestCase):
         self.func1 = node
 
     def test_implied_attrs(self):
+        func = self.func1
         decls = self.func1.ast.params
-        generate.check_implied_attrs(decls)
+        generate.check_implied_attrs(func, decls)
 
     def test_implied(self):
+        func = self.func1
         decls = self.func1.ast.params
-        expr = generate.check_implied("user(array)", decls)
+        expr = generate.check_implied(func, "user(array)", decls)
         self.assertEqual("user(array)", expr)
 
     def test_errors(self):
+        func = self.func1
         decls = self.func1.ast.params
 
         with self.assertRaises(RuntimeError) as context:
-            generate.check_implied("size(array,n2)", decls)
+            generate.check_implied(func, "size(array,n2)", decls)
         self.assertTrue("Too many arguments" in str(context.exception))
 
         with self.assertRaises(RuntimeError) as context:
-            generate.check_implied("size(array2)", decls)
+            generate.check_implied(func, "size(array2)", decls)
         self.assertTrue("Unknown argument" in str(context.exception))
 
         with self.assertRaises(RuntimeError) as context:
-            generate.check_implied("len(scalar,1)", decls)
+            generate.check_implied(func, "len(scalar,1)", decls)
         self.assertTrue("Too many arguments" in str(context.exception))